server/requirements.txt
telethon
requests
PyYAML
dnspython
dnslib
fastapi
uvicorn
